make getCooldownSeconds() return a double to the nearest tick

+++ src/main/java/com/ithinkrok/minigames/api/user/User.java
@@ -142,7 +142,7 @@ public interface User
 
     boolean isCoolingDown(String ability);
 
-    int getCooldownSeconds(String ability);
+    double getCooldownSeconds(String ability);
 
     int getUpgradeLevel(String upgrade);
